Rename aicom and features to chains and tools
