+++ mylibr/tools.py
